requirements.txt
Babel==2.4.0
blinker==1.4
cache==1.0.3
click==6.7
et-xmlfile==1.0.1
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Babel2==0.1
Flask-BabelEx==0.9.3
flask-cache-external-assets==0.2.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
healthcheck==1.3.2
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
MarkupSafe==1.0
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
passlib==1.7.1
Pillow==4.2.1
python-magic==0.4.13
pytz==2017.2
PyYAML==3.12
speaklater==1.3
SQLAlchemy==1.1.11
tablib==0.11.5
unicodecsv==0.14.1
Werkzeug==0.12.2
WTForms==2.1
xlrd==1.0.0
xlwt==1.2.0
